Se agrega configuración de seguridad y relación entre entidades. Se personaliza banner y se actualiza versión de spring boot.

+++ control-plagas-server/src/main/java/uy/edu/ude/repository/ClienteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import uy.edu.ude.entity.Cliente;
 
 @RepositoryRestResource(collectionResourceRel = "cliente", path = "cliente")
-public interface ClienteRepository extends CrudRepository<Cliente, Long> {
+public interface ClienteRepository extends CrudRepository<Cliente, Integer> {
 
 }
