+++ control-plagas-server/src/main/java/uy/edu/ude/repository/DepartamentoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import uy.edu.ude.entity.Departamento;
 
 @RepositoryRestResource(path = "departamento")
-public interface DepartamentoRepository extends PagingAndSortingRepository<Departamento, Long> {
+public interface DepartamentoRepository extends PagingAndSortingRepository<Departamento, Integer> {
 
 }
